+++ packages/mongodb-memory-server-core/src/MongoMemoryServer.ts
@@ -1,5 +1,4 @@
 import { SpawnOptions } from 'child_process';
-import * as tmp from 'tmp';
 import getPort from 'get-port';
 import {
   assertion,
@@ -10,6 +9,8 @@ import {
   statPath,
   ManagerAdvanced,
   Cleanup,
+  createTmpDir,
+  removeDir,
 } from './util/utils';
 import { MongoInstance, MongodOpts, MongoMemoryInstanceOpts } from './util/MongoInstance';
 import { MongoBinaryOpts } from './util/MongoBinary';
@@ -17,13 +18,10 @@ import debug from 'debug';
 import { EventEmitter } from 'events';
 import { promises as fspromises } from 'fs';
 import { AddUserOptions, MongoClient } from 'mongodb';
-import { lt } from 'semver';
 import { EnsureInstanceError, StateError } from './util/errors';
 import * as os from 'os';
 
 const log = debug('MongoMS:MongoMemoryServer');
-
-tmp.setGracefulCleanup();
 
 /**
  * MongoMemoryServer Stored Options
@@ -83,7 +81,7 @@ export interface StartupInstanceData {
   ip: NonNullable<MongoMemoryInstanceOpts['ip']>;
   storageEngine: NonNullable<MongoMemoryInstanceOpts['storageEngine']>;
   replSet?: NonNullable<MongoMemoryInstanceOpts['replSet']>;
-  tmpDir?: tmp.DirResult;
+  tmpDir?: string;
   keyfileLocation?: NonNullable<MongoMemoryInstanceOpts['keyfileLocation']>;
   launchTimeout?: NonNullable<MongoMemoryInstanceOpts['launchTimeout']>;
 }
@@ -229,7 +227,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
   readonly auth?: Required<AutomaticAuth>;
 
   /**
-   * Create an Mongo-Memory-Sever Instance
+   * Create a Mongo-Memory-Sever Instance
    * @param opts Mongo-Memory-Sever Options
    */
   constructor(opts?: MongoMemoryServerOpts) {
@@ -244,7 +242,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
   }
 
   /**
-   * Create an Mongo-Memory-Sever Instance that can be awaited
+   * Create a Mongo-Memory-Sever Instance that can be awaited
    * @param opts Mongo-Memory-Sever Options
    */
   static async create(opts?: MongoMemoryServerOpts): Promise<MongoMemoryServer> {
@@ -330,13 +328,13 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
   }
 
   /**
-   * Find an new unlocked port
-   * @param port An User defined default port
+   * Find a new unlocked port
+   * @param port A User defined default port
    */
   protected async getNewPort(port?: number): Promise<number> {
     const newPort = await getPort({ port });
 
-    // only log this message if an custom port was provided
+    // only log this message if a custom port was provided
     if (port != newPort && typeof port === 'number') {
       this.debug(`getNewPort: starting with port "${newPort}", since "${port}" was locked`);
     }
@@ -380,16 +378,12 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
     };
 
     if (isNullOrUndefined(this._instanceInfo)) {
-      // create an tmpDir instance if no "dbPath" is given
+      // create a tmpDir instance if no "dbPath" is given
       if (!data.dbPath) {
-        data.tmpDir = tmp.dirSync({
-          mode: 0o755,
-          prefix: 'mongo-mem-',
-          unsafeCleanup: true,
-        });
-        data.dbPath = data.tmpDir.name;
+        data.tmpDir = await createTmpDir('mongo-mem-');
+        data.dbPath = data.tmpDir;
 
-        isNew = true; // just to ensure "isNew" is "true" because an new temporary directory got created
+        isNew = true; // just to ensure "isNew" is "true" because a new temporary directory got created
       } else {
         this.debug(
           `getStartOptions: Checking if "${data.dbPath}}" (no new tmpDir) already has data`
@@ -410,7 +404,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
       enableAuth && // re-use all the checks from "enableAuth"
       !isNullOrUndefined(this.auth) && // needs to be re-checked because typescript complains
       (this.auth.force || isNew) && // check that either "isNew" or "this.auth.force" is "true"
-      !instOpts.replSet; // dont run "createAuth" when its an replset, it will be run by the replset controller
+      !instOpts.replSet; // dont run "createAuth" when its a replset, it will be run by the replset controller
 
     return {
       data: data,
@@ -603,8 +597,8 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
     const tmpDir = this._instanceInfo.tmpDir;
 
     if (!isNullOrUndefined(tmpDir)) {
-      this.debug(`cleanup: removing tmpDir at ${tmpDir.name}`);
-      tmpDir.removeCallback();
+      this.debug(`cleanup: removing tmpDir at ${tmpDir}`);
+      await removeDir(tmpDir);
     }
 
     if (cleanup.force) {
@@ -614,15 +608,9 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
       if (isNullOrUndefined(res)) {
         this.debug(`cleanup: force is true, but path "${dbPath}" dosnt exist anymore`);
       } else {
-        assertion(res.isDirectory(), new Error('Defined dbPath is not an directory'));
+        assertion(res.isDirectory(), new Error('Defined dbPath is not a directory'));
 
-        if (lt(process.version, '14.14.0')) {
-          // this has to be used for 12.10 - 14.13 (inclusive) because ".rm" did not exist yet
-          await fspromises.rmdir(dbPath, { recursive: true, maxRetries: 1 });
-        } else {
-          // this has to be used for 14.14+ (inclusive) because ".rmdir" and "recursive" got deprecated (DEP0147)
-          await fspromises.rm(dbPath, { recursive: true, maxRetries: 1 });
-        }
+        await removeDir(dbPath);
       }
     }
 
@@ -667,7 +655,7 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
             if (state != MongoMemoryServerStates.running) {
               rej(
                 new Error(
-                  `"ensureInstance" waited for "running" but got an different state: "${state}"`
+                  `"ensureInstance" waited for "running" but got a different state: "${state}"`
                 )
               );
 
@@ -709,11 +697,11 @@ export class MongoMemoryServer extends EventEmitter implements ManagerAdvanced {
 
   /**
    * Generate the Connection string used by mongodb
-   * @param otherDb add an database into the uri (in mongodb its the auth database, in mongoose its the default database for models)
+   * @param otherDb add a database into the uri (in mongodb its the auth database, in mongoose its the default database for models)
    * @param otherIp change the ip in the generated uri, default will otherwise always be "127.0.0.1"
    * @throws if state is not "running" (or "starting")
-   * @throws if an server doesnt have "instanceInfo.port" defined
-   * @returns an valid mongo URI, by the definition of https://docs.mongodb.com/manual/reference/connection-string/
+   * @throws if a server doesnt have "instanceInfo.port" defined
+   * @returns a valid mongo URI, by the definition of https://docs.mongodb.com/manual/reference/connection-string/
    */
   getUri(otherDb?: string, otherIp?: string): string {
     this.debug('getUri:', this.state, otherDb, otherIp);