+++ packages/mongodb-memory-server-core/src/__tests__/MongoMemoryServer.test.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { MongoClient, MongoServerError } from 'mongodb';
-import * as tmp from 'tmp';
 import MongoMemoryServer, {
   CreateUser,
   MongoInstanceData,
@@ -9,13 +8,11 @@ import MongoMemoryServer, {
 } from '../MongoMemoryServer';
 import MongoInstance from '../util/MongoInstance';
 import * as utils from '../util/utils';
-import * as semver from 'semver';
 import { EnsureInstanceError, StateError } from '../util/errors';
 import { assertIsError } from './testUtils/test_utils';
 import { promises as fspromises } from 'fs';
 import * as path from 'path';
 
-tmp.setGracefulCleanup();
 jest.setTimeout(100000); // 10s
 
 afterEach(() => {
@@ -581,7 +578,7 @@ describe('MongoMemoryServer', () => {
       mongoServer.emit(MongoMemoryServerEvents.stateChange, MongoMemoryServerStates.stopped);
 
       expect(ensureInstancePromise).rejects.toThrow(
-        `"ensureInstance" waited for "running" but got an different state: "${MongoMemoryServerStates.stopped}"`
+        `"ensureInstance" waited for "running" but got a different state: "${MongoMemoryServerStates.stopped}"`
       );
     });
 
@@ -770,18 +767,17 @@ describe('MongoMemoryServer', () => {
 
   describe('cleanup()', () => {
     /** Cleanup the created tmp-dir, even if the cleanup test tested to not clean it up */
-    let tmpdir: tmp.DirResult | undefined;
+    let tmpdir: string | undefined;
 
     // "beforeAll" dosnt work here, thanks to the top-level "afterAll" hook
     beforeEach(() => {
       jest.spyOn(utils, 'statPath');
-      // @ts-expect-error because "default" dosnt exist in the definitions
-      jest.spyOn(semver.default, 'lt'); // it needs to be ".default" otherwise "lt" is only an getter
+      jest.spyOn(utils, 'removeDir');
     });
 
-    afterEach(() => {
+    afterEach(async () => {
       if (!utils.isNullOrUndefined(tmpdir)) {
-        tmpdir.removeCallback();
+        await utils.removeDir(tmpdir);
 
         tmpdir = undefined; // reset, just to be sure its clean
       }
@@ -797,7 +793,7 @@ describe('MongoMemoryServer', () => {
       await mongoServer.cleanup();
 
       expect(utils.statPath).not.toHaveBeenCalled();
-      expect(semver.lt).not.toHaveBeenCalled();
+      expect(utils.removeDir).toHaveBeenCalled();
       expect(await utils.statPath(dbPath)).toBeUndefined();
       expect(mongoServer.state).toEqual(MongoMemoryServerStates.new);
       expect(mongoServer.instanceInfo).toBeUndefined();
@@ -813,15 +809,15 @@ describe('MongoMemoryServer', () => {
       await mongoServer.cleanup(true);
 
       expect(utils.statPath).toHaveBeenCalledTimes(1);
-      expect(semver.lt).not.toHaveBeenCalled();
+      expect(utils.removeDir).toHaveBeenCalled();
       expect(await utils.statPath(dbPath)).toBeUndefined();
       expect(mongoServer.state).toEqual(MongoMemoryServerStates.new);
       expect(mongoServer.instanceInfo).toBeUndefined();
     });
 
     it('should properly cleanup with force (without tmpDir) (old)', async () => {
-      const tmpDir = tmp.dirSync({ prefix: 'mongo-mem-cleanup-', unsafeCleanup: true });
-      const mongoServer = await MongoMemoryServer.create({ instance: { dbPath: tmpDir.name } });
+      const tmpDir = await utils.createTmpDir('mongo-mem-cleanup-');
+      const mongoServer = await MongoMemoryServer.create({ instance: { dbPath: tmpDir } });
       const dbPath = mongoServer.instanceInfo!.dbPath;
 
       tmpdir = mongoServer.instanceInfo?.tmpDir;
@@ -830,7 +826,7 @@ describe('MongoMemoryServer', () => {
       await mongoServer.cleanup(true);
 
       expect(utils.statPath).toHaveBeenCalledTimes(1);
-      expect(semver.lt).toHaveBeenCalled(); // not testing on how many, because it would change with nodejs version
+      expect(utils.removeDir).toHaveBeenCalled();
       expect(await utils.statPath(dbPath)).toBeUndefined();
       expect(mongoServer.state).toEqual(MongoMemoryServerStates.new);
       expect(mongoServer.instanceInfo).toBeUndefined();
@@ -847,7 +843,7 @@ describe('MongoMemoryServer', () => {
       await mongoServer.stop({ doCleanup: false });
       await mongoServer.cleanup();
       expect(utils.statPath).not.toHaveBeenCalled();
-      expect(semver.lt).not.toHaveBeenCalled();
+      expect(utils.removeDir).toHaveBeenCalled();
       expect(await utils.statPath(dbPath)).toBeUndefined();
       expect(mongoServer.state).toEqual(MongoMemoryServerStates.new);
       expect(mongoServer.instanceInfo).toBeUndefined();
@@ -865,7 +861,7 @@ describe('MongoMemoryServer', () => {
       await mongoServer.stop({ doCleanup: false });
       await mongoServer.cleanup({ doCleanup: true, force: true });
       expect(utils.statPath).toHaveBeenCalledTimes(1);
-      expect(semver.lt).not.toHaveBeenCalled();
+      expect(utils.removeDir).toHaveBeenCalled();
       expect(await utils.statPath(dbPath)).toBeUndefined();
       expect(mongoServer.state).toEqual(MongoMemoryServerStates.new);
       expect(mongoServer.instanceInfo).toBeUndefined();
@@ -873,8 +869,8 @@ describe('MongoMemoryServer', () => {
     });
 
     it('should properly cleanup with force (without tmpDir) (new)', async () => {
-      const tmpDir = tmp.dirSync({ prefix: 'mongo-mem-cleanup-', unsafeCleanup: true });
-      const mongoServer = await MongoMemoryServer.create({ instance: { dbPath: tmpDir.name } });
+      const tmpDir = await utils.createTmpDir('mongo-mem-cleanup-');
+      const mongoServer = await MongoMemoryServer.create({ instance: { dbPath: tmpDir } });
       const dbPath = mongoServer.instanceInfo!.dbPath;
 
       const cleanupSpy = jest.spyOn(mongoServer, 'cleanup');
@@ -884,7 +880,7 @@ describe('MongoMemoryServer', () => {
       await mongoServer.stop({ doCleanup: false });
       await mongoServer.cleanup({ doCleanup: true, force: true });
       expect(utils.statPath).toHaveBeenCalledTimes(1);
-      expect(semver.lt).toHaveBeenCalled(); // not testing on how many, because it would change with nodejs version
+      expect(utils.removeDir).toHaveBeenCalled();
       expect(await utils.statPath(dbPath)).toBeUndefined();
       expect(mongoServer.state).toEqual(MongoMemoryServerStates.new);
       expect(mongoServer.instanceInfo).toBeUndefined();
@@ -926,10 +922,10 @@ describe('MongoMemoryServer', () => {
 
     it('should resolve "isNew" to "true" and set "createAuth" to "true" when dbPath is set, but empty', async () => {
       const readdirSpy = jest.spyOn(fspromises, 'readdir');
-      const tmpDbPath = tmp.dirSync({ prefix: 'mongo-mem-getStartOptions1-', unsafeCleanup: true });
+      const tmpDbPath = await utils.createTmpDir('mongo-mem-getStartOptions1-');
 
       const mongoServer = new MongoMemoryServer({
-        instance: { dbPath: tmpDbPath.name },
+        instance: { dbPath: tmpDbPath },
         auth: {},
       });
 
@@ -942,22 +938,22 @@ describe('MongoMemoryServer', () => {
         new Error('Expected "options.data.dbPath" to be defined')
       );
       expect(await utils.pathExists(options.data.dbPath)).toEqual(true);
-      expect(options.data.dbPath).toEqual(tmpDbPath.name);
+      expect(options.data.dbPath).toEqual(tmpDbPath);
       expect(readdirSpy).toHaveBeenCalledTimes(1);
       expect(options.createAuth).toEqual(true);
 
-      tmpDbPath.removeCallback();
+      await utils.removeDir(tmpDbPath);
     });
 
     it('should resolve "isNew" to "false" and set "createAuth" to "false" when dbPath is set, but not empty', async () => {
       const readdirSpy = jest.spyOn(fspromises, 'readdir');
-      const tmpDbPath = tmp.dirSync({ prefix: 'mongo-mem-getStartOptions1-', unsafeCleanup: true });
+      const tmpDbPath = await utils.createTmpDir('mongo-mem-getStartOptions2-');
 
       // create dummy file, to make the directory non-empty
-      await fspromises.writeFile(path.resolve(tmpDbPath.name, 'testfile'), '');
+      await fspromises.writeFile(path.resolve(tmpDbPath, 'testfile'), '');
 
       const mongoServer = new MongoMemoryServer({
-        instance: { dbPath: tmpDbPath.name },
+        instance: { dbPath: tmpDbPath },
         auth: {},
       });
 
@@ -970,11 +966,11 @@ describe('MongoMemoryServer', () => {
         new Error('Expected "options.data.dbPath" to be defined')
       );
       expect(await utils.pathExists(options.data.dbPath)).toEqual(true);
-      expect(options.data.dbPath).toEqual(tmpDbPath.name);
+      expect(options.data.dbPath).toEqual(tmpDbPath);
       expect(readdirSpy).toHaveBeenCalledTimes(1);
       expect(options.createAuth).toEqual(false);
 
-      tmpDbPath.removeCallback();
+      await utils.removeDir(tmpDbPath);
     });
 
     it('should generate a port when no suggestion is defined', async () => {
@@ -1027,17 +1023,17 @@ describe('MongoMemoryServer', () => {
   });
 
   it('"getDbPath" should return the dbPath', async () => {
-    const tmpDir = tmp.dirSync({ prefix: 'mongo-mem-getDbPath-', unsafeCleanup: true });
+    const tmpDir = await utils.createTmpDir('mongo-mem-getDbPath-');
     const mongoServer = new MongoMemoryServer({
-      instance: { dbPath: tmpDir.name },
+      instance: { dbPath: tmpDir },
     });
 
     await mongoServer.start();
 
-    expect(mongoServer.instanceInfo!.dbPath).toEqual(tmpDir.name);
+    expect(mongoServer.instanceInfo!.dbPath).toEqual(tmpDir);
 
     await mongoServer.stop();
-    tmpDir.removeCallback();
+    await utils.removeDir(tmpDir);
   });
 
   it('"state" should return correct state', () => {